refactor message classes
