+++ lib/src/messages/raygun_app_context.dart
@@ -1,0 +1,8 @@
+import 'package:json_annotation/json_annotation.dart';
+
+part 'raygun_app_context.g.dart';
+
+@JsonSerializable()
+class RaygunAppContext {
+  Map<String, dynamic> toJson() => _$RaygunAppContextToJson(this);
+}
