+++ lib/src/messages/raygun_breadcrumb_message.dart
