+++ lib/src/messages/raygun_error_stack_trace_line_message.dart
@@ -1,0 +1,21 @@
+import 'package:json_annotation/json_annotation.dart';
+
+part 'raygun_error_stack_trace_line_message.g.dart';
+
+@JsonSerializable()
+class RaygunErrorStackTraceLineMessage {
+  final int? lineNumber;
+  final String? className;
+  final String? fileName;
+  final String? methodName;
+
+  RaygunErrorStackTraceLineMessage(
+    this.lineNumber,
+    this.className,
+    this.fileName,
+    this.methodName,
+  );
+
+  Map<String, dynamic> toJson() =>
+      _$RaygunErrorStackTraceLineMessageToJson(this);
+}
