+++ lib/src/messages/raygun_message.dart
@@ -1,0 +1,18 @@
+import 'package:json_annotation/json_annotation.dart';
+
+import 'raygun_message_details.dart';
+
+part 'raygun_message.g.dart';
+
+@JsonSerializable()
+class RaygunMessage {
+  late String occurredOn;
+  late RaygunMessageDetails details;
+
+  RaygunMessage() {
+    occurredOn = 'TODO DATE';
+    details = RaygunMessageDetails();
+  }
+
+  Map<String, dynamic> toJson() => _$RaygunMessageToJson(this);
+}
