+++ lib/src/raygun_crash_reporting.dart
@@ -6,7 +6,10 @@ import 'package:device_info_plus/device_info_plus.dart';
 import 'package:stack_trace/stack_trace.dart';
 
 import 'logging/raygun_logger.dart';
-import 'messages/raygun_messages.dart';
+import 'messages/raygun_client_message.dart';
+import 'messages/raygun_environment_message.dart';
+import 'messages/raygun_error_message.dart';
+import 'messages/raygun_message.dart';
 
 class CrashReporting {
   static Future<void> send(
